bugfix: Fix NPE that sometimes occurs with zipped roms.
Fixes crash
https://paulscode.iriscouch.com/acralyzer-custom/_design/acralyzer/index.html#/reports-browser/mupen64plusae/bug/d16a63ff17b04dffdcb2bee74a51fd66

+++ src/paulscode/android/mupen64plusae/CoreInterface.java
@@ -546,9 +546,12 @@ public class CoreInterface
             
             // Clear the folder if anything is in there:
             String[] children = tmpFolder.list();
-            for( String child : children )
+            if( children != null )
             {
-                FileUtil.deleteFolder( new File( tmpFolder, child ) );
+                for( String child : children )
+                {
+                    FileUtil.deleteFolder( new File( tmpFolder, child ) );
+                }
             }
             
             // Unzip the ROM
